homework/requirements.txt
asgiref==3.3.4
attrs==21.2.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==4.4.0
iniconfig==1.1.1
Markdown==3.3.4
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.4.1
